requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
dart-fss
requests